--- a/src/algo/spfa.rs
+++ b/src/algo/spfa.rs
@@ -1,4 +1,6 @@
 //! Shortest Path Faster Algorithm.
+use alloc::collections::VecDeque;
+
 use super::{bellman_ford::Paths, BoundedMeasure, NegativeCycle};
 use crate::prelude::*;
 use crate::visit::{IntoEdges, IntoNodeIdentifiers, NodeIndexable};
@@ -93,10 +95,10 @@ where
     dist[ix(source)] = K::default();
 
     // Queue of vertices capable of relaxation of the found shortest distances.
-    let mut queue: Vec<G::NodeId> = Vec::with_capacity(graph.node_bound());
+    let mut queue: VecDeque<G::NodeId> = VecDeque::with_capacity(graph.node_bound());
     let mut in_queue = vec![false; graph.node_bound()];
 
-    queue.push(source);
+    queue.push_back(source);
     in_queue[ix(source)] = true;
 
     let (distances, predecessors) = spfa_loop(graph, dist, Some(pred), queue, in_queue, edge_cost)?;
@@ -116,7 +118,7 @@ pub(crate) fn spfa_loop<G, F, K>(
     graph: G,
     mut distances: Vec<K>,
     mut predecessors: Option<Vec<Option<G::NodeId>>>,
-    mut queue: Vec<G::NodeId>,
+    mut queue: VecDeque<G::NodeId>,
     mut in_queue: Vec<bool>,
     mut edge_cost: F,
 ) -> Result<(Vec<K>, Option<Vec<Option<G::NodeId>>>), NegativeCycle>
@@ -131,7 +133,7 @@ where
     // in the queue to be able to detect a negative cycle.
     let mut visits = vec![0; graph.node_bound()];
 
-    while let Some(i) = queue.pop() {
+    while let Some(i) = queue.pop_front() {
         in_queue[ix(i)] = false;
 
         // In a graph without a negative cycle, no vertex can improve
@@ -155,7 +157,7 @@ where
 
                 if !in_queue[ix(j)] {
                     in_queue[ix(j)] = true;
-                    queue.push(j);
+                    queue.push_back(j);
                 }
             }
         }