<!--
  -- Thanks for contributing to `petgraph`! 
  --
  -- We require PR titles to follow the Conventional Commits specification,
  -- https://www.conventionalcommits.org/en/v1.0.0/. This helps us generate
  -- changelogs and follow semantic versioning.
  --
  -- Start the PR title with one of the following:
  --  * `feat:` for new features
  --  * `fix:` for bug fixes
  --  * `refactor:` for code refactors
  --  * `docs:` for documentation changes
  --  * `test:` for test changes
  --  * `perf:` for performance improvements
  --  * `revert:` for reverting changes
  --  * `ci:` for CI/CD changes
  --  * `chore:` for changes that don't fit in any of the above categories
  -- The last two categories will not be included in the changelog.
  --
  -- If your PR includes a breaking change, please add a `!` after the type
  -- and include a `BREAKING CHANGE:` line in the body of the PR describing
  -- the necessary changes for users to update their code.
  --
  -->
  
fixes https://github.com/petgraph/petgraph/issues/862

The stack-based implementation of SPFA produces false negatives when detecting negative cycles. When benchmarking without a negative cycle, it actually runs faster than Bellman-Ford.

Since the SPFA implementation uses a fixed-size(capacity) queue, no reallocation occurs in `VecDeque`, so it is expected to be sufficiently fast.

**Before**: 46082d680ac1b56ba19da61ec90572307480eb27
```
test bellman_ford_bench        ... bench:       1,813.35 ns/iter (+/- 81.06)
test spfa_bench                ... bench:      21,197.04 ns/iter (+/- 1,145.72)
```
  
**After**: 10051b28c81c0bf5c7f32ba0be3fc5e391b84303
```
test bellman_ford_bench        ... bench:       1,826.56 ns/iter (+/- 155.06)
test spfa_bench                ... bench:       1,371.46 ns/iter (+/- 110.51)
```