@RaoulLuque 2025-09-28
Looks great, thank you so much :+1: Would you mind rebasing and moving the test into `tests/spfa.rs`?